requirements/base.txt
Flask==3.0.2
python-dotenv==1.0.1
pre-commit==3.3.3
Flask-Cors==3.0.10
psycopg2-binary==2.9.6
camelot-py==0.9.0
numpy==1.26.4
opencv-python==4.9.0.80
PyPDF2==2.12.1
tabula-py==2.9.0
pytest==8.0.0
flask-swagger==0.2.14
flask_restx==1.3.0